init commit

+++ src/main/java/web/сontroller/AdminController.java
@@ -1,16 +1,22 @@
 package web.сontroller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.*;
+import web.model.Role;
 import web.model.User;
 import web.service.RoleService;
 import web.service.UserService;
 
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
-@Controller
+@RestController
 @RequestMapping("/admin")
 public class AdminController {
 
@@ -23,30 +29,58 @@ public class AdminController {
         this.roleService = roleService;
     }
 
-    @GetMapping("/")
-    public String showAllUsers(ModelMap model, @AuthenticationPrincipal User user) {
-        model.addAttribute("currentUser", user);
-        model.addAttribute("listUsers", userService.findAll());
-        model.addAttribute("allRoles", roleService.findAll());
-        return "users";
+    public AdminController() {
     }
 
-    @PostMapping("/")
-    public String addUser(@ModelAttribute("user") User user, @RequestParam("chosenRoles") Long[] roles) {
-        userService.save(user, roles);
-        return "redirect:/admin/";
+    @GetMapping("/users/{id}")
+    public ResponseEntity<User> getUser(@PathVariable("id") Long userId) {
+        if (userId == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        Optional<User> user = userService.findById(userId);
+        return user.map(u -> new ResponseEntity<>(u, HttpStatus.OK))
+                .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
-    @PostMapping("/update")
-    public String updateUser(User user,@RequestParam("id") Long id, @RequestParam("chosenRoles") Long[] roles) {
-        user.setId(id);
-        userService.update(user, roles);
-        return "redirect:/admin/";
+    @PostMapping("/users")
+    public ResponseEntity<User> saveUser(@RequestBody User user) {
+        this.userService.save(user);
+        return ResponseEntity.ok(user);
     }
 
-    @GetMapping(value = "/delete")
-    public String deleteUsers(@RequestParam("id") int id) {
-        userService.deleteById(id);
-        return "redirect:/admin/";
+    @PutMapping("/users")
+    public ResponseEntity<User> updateUser(@RequestBody User user) {
+        userService.update(user);
+        return ResponseEntity.ok(user);
+    }
+
+    @DeleteMapping("/users/{id}")
+    public ResponseEntity<User> deleteUser(@PathVariable("id") Long id) {
+        if (userService.findById(id).isPresent()) {
+            userService.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @GetMapping("/users")
+    public ResponseEntity<List<User>> getAllUsers() {
+        List<User> users = userService.findAll();
+
+        if (users.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(users, HttpStatus.OK);
+    }
+
+    @GetMapping("/roles")
+    public ResponseEntity<List<Role>> getAllRoles() {
+        List<Role> roles = this.roleService.findAll();
+
+        if (roles.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(roles, HttpStatus.OK);
     }
 }
